requirements.txt
# DocQuiz — requirements.txt
# Install with: pip install -r requirements.txt

# UI
streamlit>=1.35.0

# Document parsing
pymupdf>=1.24.0          # PDF  — import as fitz
python-docx>=1.1.0       # DOCX

# HTTP (Ollama API calls)
requests>=2.31.0

# Utilities (usually already present, listed for reproducibility)
pathlib2>=2.3.7; python_version < "3.4"
